fix cookies for sorting the table
